[황상욱] 덫 밟으면 스턴 이펙트, 조작시 카메라 움직임 정상확인

+++ CardDungeon/Assets/HJH/Script/PlayerDeck_HJH.cs
@@ -378,12 +378,12 @@ public class PlayerDeck_HJH : MonoBehaviour
     }
     #endregion
 
-    public void StunGo(int time)
+    public void StunGo(float time)
     {
         StartCoroutine(Stun(time));
     }
 
-    IEnumerator Stun(int time)
+    IEnumerator Stun(float time)
     {
         stun = true;
         yield return new WaitForSeconds(time);
